Finish writing decompiler
